Add copy to clipboard when getting a value
Copy to clipboard is a superb feature in the original BOOM which when you
look for `boom.net <bucket> <key>`, you get back a value which is ALSO
copied to the clipboard. This making storing URLs in BOOM super nifty

Adds the Windows.Forms dependency to make use of the clipboard and needs
to annotate the main function with `STAThread` to ensure some COM
interop happens properly since COM is a single threaded model.

Read more here:
http://blogs.msdn.com/b/jfoscoding/archive/2005/04/07/406341.aspx