Start to make app work on .NET framework also
This requires conditionals to work. the `#if DNXCORE50` is a
preprocessor directive that can be used to check if a set of code has to
only be executed if it is run on the coreclr.

Moving the OS check to a `Util` library which does this check and
accordingly returns the value.